+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 
 ### Changed
 
+- Template management widget delete button now requires a confirmation before deleting a template.
 - Sandbox changes:
   - Rebuilt the input for sandbox, it is not a _mini wizard_ that works very similarly to the normal work flow. For series we needed more widgets to gather input from the user, so this allows a nice easy to use flow for the user.
   - When the user is opening a series, the user will now see **Series Episode Matcher** page, to allow them to confirm/match their episodes with TVDB data.